Inference works, but reults in split graph
